# Add role tags to user info in flex tab