--- a/packages/rocketchat-theme/assets/stylesheets/base.less
+++ b/packages/rocketchat-theme/assets/stylesheets/base.less
@@ -3160,6 +3160,12 @@ body:not(.is-cordova) {
 			font-size: 12px;
 			font-weight: 300;
 		}
+		.role-tag {
+			background-color: #ccc;
+			padding: 1px 4px;
+			border-radius: 2px;
+			color: white;
+		}
 	}
 	.stats {
 		li {