--- a/packages/rocketchat-ui-flextab/flex-tab/tabs/userInfo.html
+++ b/packages/rocketchat-ui-flextab/flex-tab/tabs/userInfo.html
@@ -13,6 +13,11 @@
 				<div class="info">
 					<h3 title="{{username}}"><i class="status-{{status}}"></i> {{username}}</h3>
 					<p>{{name}}</p>
+					<p>
+						{{#each roleTags}}
+							<span class="role-tag" data-role="{{.}}">{{.}}</span>
+						{{/each}}
+					</p>
 					{{#if utc}}<p><i class="icon-clock"></i>{{userTime}} (UTC {{utc}})</p>{{/if}}
 					{{#if hasPermission 'view-full-other-user-info'}}
 						{{#if hasEmails}}