--- a/packages/rocketchat-ui-message/message/message.html
+++ b/packages/rocketchat-ui-message/message/message.html
@@ -28,7 +28,7 @@
 		{{/if}}
 		<span class="info">
 			{{#each roleTags}}
-				<span class="role-tag" data-role="{{.}}">{{.}}</span>
+				<span class="role-tag" data-role="{{description}}">{{description}}</span>
 			{{/each}}
 			{{#if isBot}}
 				<span class="is-bot">BOT</span>