@RocketChat/core 

![ss](https://cloud.githubusercontent.com/assets/9736483/15453305/ec5c1454-204b-11e6-95c2-213a1b680487.png)