@lucasvanhalst 2016-05-22
Doesn't the `roleTags` function execute like 10 queries if the user has 5 roles assigned?