@nishimaki10 2016-05-22
@lucasvanhalst Certainly it looks like you said. However, I see [this](https://github.com/RocketChat/Rocket.Chat/blob/develop/packages/rocketchat-ui-message/message/message.coffee#L4).
Hmmm... I will try to improve. Thanks.