@lucasvanhalst 2016-05-22
I guess you could use an [$in](https://docs.mongodb.com/v3.0/reference/operator/query/in/) query, like this: 
```javascript
model.find({
    _id: {
        $in: roles
    },
    description: {
        $exists: 1,
        $ne: ''
    }
}, {
    description: 1
})
```