@nishimaki10 2016-05-22
Yes, thanks. I finished.
I'm no using `$ne: ''`, but look like no problem. Empty role tag element does not generated.