@sampaiodiego 2016-05-23
Hey guys, I'm glad to see a performance concern from both of you =)
just for reference, I made [this test code](https://gist.github.com/sampaiodiego/8e559988093989a9cde1bce411cb3929) to compare the performance of the "roles tags selector" and got the following results:

  | old code (ms) | new code (ms)
----|--------|------
1 role | 851 | 1763
2 roles | 1951 | 2294
3 roles | 2833 | 2032
4 roles | 4105 | 2202

I'm looking into this PR and the from @lucasvanhalst .. hope to merge both soon :wink: 

good job guys.. thanks for the contribution