@lucasvanhalst 2016-05-23
@sampaiodiego That's interesting.
Then again, I thought this was a database query being done because I didn't quite understand the differences between client and server, so after all it doesn't really matter I guess.