@sampaiodiego 2016-05-23
@lucasvanhalst it's only client side.. but I liked the improvement.. it's more consistent no matter the number of roles a user has. =)